pyproject.toml
[project]
name = "cash-register"
version = "0.1.0"
description = "Test task project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.10",
    "pdfkit>=1.0.0",
    "pillow>=11.1.0",
    "qrcode>=8.0",
    "requests>=2.32.3",
    "wkhtmltopdf>=0.2",
]
